Put checklist under student profile, fix routes
