+++ app/account/views.py
@@ -9,7 +9,7 @@ from ..email import send_email
 from ..models import User, ChecklistItem
 from .forms import (ChangeEmailForm, ChangePasswordForm, CreatePasswordForm,
                     LoginForm, RegistrationForm, RequestResetPasswordForm,
-                    ResetPasswordForm, AddChecklistItemForm, EditChecklistItemForm)
+                    ResetPasswordForm)
 
 
 @account.route('/login', methods=['GET', 'POST'])
@@ -271,77 +271,3 @@ def unconfirmed():
     if current_user.is_anonymous or current_user.confirmed:
         return redirect(url_for('main.index'))
     return render_template('account/unconfirmed.html')
-
-
-@account.route('/checklist', methods=['GET', 'POST'])
-@login_required
-def checklist():
-    checklist_items = ChecklistItem.query.filter_by(assignee_id=current_user.id)
-    completed_items = [item for item in checklist_items if item.is_checked]
-    checklist_items = [item for item in checklist_items if not item.is_checked]
-    form = AddChecklistItemForm()
-    if form.validate_on_submit():
-        #add new checklist item to user's account
-        checklist_item = ChecklistItem(
-            assignee_id=current_user.id,
-            text=form.item_text.data,
-            is_deletable=True)
-        db.session.add(checklist_item)
-        db.session.commit()
-        return redirect(url_for('account.checklist'))
-    return render_template('account/checklist.html', form=form, checklist=checklist_items, 
-        completed=completed_items)
-
-
-@account.route('/checklist/delete/<int:item_id>', methods=['GET', 'POST'])
-@login_required
-def delete_checklist_item(item_id):
-    checklist_item = ChecklistItem.query.filter_by(id=item_id).first()
-    if checklist_item:
-        db.session.delete(checklist_item)
-        db.session.commit()
-        return redirect(url_for('account.checklist'))
-    flash('Item could not be deleted', 'error')
-    return redirect(url_for('account.checklist'))
-
-
-@account.route('/checklist/complete/<int:item_id>', methods=['GET', 'POST'])
-@login_required
-def complete_checklist_item(item_id):
-    checklist_item = ChecklistItem.query.filter_by(id=item_id).first()
-    if checklist_item:
-        checklist_item.is_checked = True
-        db.session.add(checklist_item)
-        db.session.commit()
-        return redirect(url_for('account.checklist'))
-    flash('Item could not be completed', 'error')
-    return redirect(url_for('account.checklist'))
-
-@account.route('/checklist/undo/<int:item_id>', methods=['GET', 'POST'])
-@login_required
-def undo_checklist_item(item_id):
-    checklist_item = ChecklistItem.query.filter_by(id=item_id).first()
-    if checklist_item:
-        checklist_item.is_checked = False
-        db.session.add(checklist_item)
-        db.session.commit()
-        return redirect(url_for('account.checklist'))
-    flash('Item could not be undone', 'error')
-    return redirect(url_for('account.checklist'))
-
-
-@account.route('/checklist/update/<int:item_id>', methods=['GET', 'POST'])
-@login_required
-def update_checklist_item(item_id):
-    item = ChecklistItem.query.filter_by(id=item_id).first()
-    if item:
-        form = EditChecklistItemForm(item_text=item.text)
-        if form.validate_on_submit():
-            #update checklist item's text
-            item.text=form.item_text.data
-            db.session.add(item)
-            db.session.commit()
-            return redirect(url_for('account.checklist'))
-        return render_template('account/update_checklist.html', form=form)
-    flash('Item could not be updated', 'error')
-    return redirect(url_for('account.checklist'))
