+++ app/models/user.py
@@ -4,7 +4,7 @@ from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 from itsdangerous import BadSignature, SignatureExpired
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from . import StudentProfile, ChecklistItem
+from . import StudentProfile
 from .. import db, login_manager
 
 
@@ -68,7 +68,6 @@ class User(UserMixin, db.Model):
         "StudentProfile",
         uselist=False,
         back_populates="user")
-    checklist = db.relationship('ChecklistItem', back_populates='assignee')
 
     def __init__(self, **kwargs):
         super(User, self).__init__(**kwargs)
@@ -196,7 +195,6 @@ class User(UserMixin, db.Model):
                 password=fake.password(),
                 confirmed=True,
                 role=role,
-                checklist = ChecklistItem.generate_fake(),
                 phone_number=fake.phone_number(),
                 **kwargs)
             if role.name == 'User':
