+++ app/templates/account/update_checklist.html
@@ -5,6 +5,6 @@
 
 	{{ f.render_form(form, extra_classes='fluid') }}
 
-    <a class="ui small basic icon button" href="{{ url_for('account.checklist') }}">
+    <a class="ui small basic icon button" href="{{ url_for('student.checklist') }}">
     Back</a>
 {% endblock %}
