+++ app/templates/student/checklist.html
@@ -1,8 +1,66 @@
-{% import 'macros/page_macros.html' as page %}
 {% extends 'layouts/base.html' %}
+{% import 'macros/form_macros.html' as f %}
+
+{% macro render_checklist(checklist) %}
+    <table class="ui basic table">
+    	{% for item in checklist %}
+            <tr>
+            	<td>{{ '%s' % (item.text) }}</td> 
+            	<td class = "right aligned">
+
+                    <a class="ui small basic icon button" href="{{ url_for('student.complete_checklist_item', item_id=item.id) }}">
+                        <i class="checkmark box icon"></i>
+                    </a>
+
+                    {% if item.is_deletable %}
+
+                        <a class="ui small basic icon button" href="{{ url_for('student.update_checklist_item', item_id=item.id) }}">
+                            <i class="write icon"></i>
+                        </a>
+                    
+                        <a class="ui small basic icon button" href="{{ url_for('student.delete_checklist_item', item_id=item.id) }}">
+                            <i class="remove icon"></i>
+                        </a>
+                    {% endif %}
+            	</td>
+            </tr>
+        {% endfor %}
+    </table>
+{% endmacro %}
+
+{% macro completed_list(completed) %}
+    <table class="ui basic table">
+        {% for item in completed %}
+            <tr>
+                <td>{{ '%s' % (item.text) }}</td> 
+                <td class = "right aligned">
+                    {% if item.is_deletable %}
+                        <a class="ui small basic icon button" href="{{ url_for('student.delete_checklist_item', item_id=item.id) }}">
+                            <i class="remove icon"></i>
+                        </a>
+                    {% endif %}
+                    <a class="ui small basic icon button" href="{{ url_for('student.undo_checklist_item', item_id=item.id) }}">
+                            <i class="undo icon"></i>
+                        </a>
+                </td>
+            </tr>
+        {% endfor %}
+    </table>
+{% endmacro %}
 
 {% block content %}
-    <div class="ui text container">
-        <h1>Checklist</h1>
+    <div class="ui stackable centered grid container">
+        <div class="twelve wide column">
+            <h2 class="ui header">
+                Checklist
+            </h2>
+        </div>
+        <div class="stretched divided very relaxed row">
+            <div class="eight wide column">
+            	{{ render_checklist(checklist) }}
+                {{ f.render_form(form, extra_classes='fluid') }}
+                {{ completed_list(completed) }}
+            </div>
+        </div>
     </div>
 {% endblock %}
